requirements.txt
RPi.GPIO==0.7.1
schedule==1.2.0
python-dotenv==1.0.0
